+++ src/api/models.py
@@ -150,7 +150,7 @@ class Landlords(db.Model):
     name = db.Column(db.String(), nullable=True)
     lastname = db.Column(db.String(), nullable=True)
     birth_date = db.Column(db.Date, nullable=True)
-    dni = db.Column(db.String(), unique=True)
+    dni = db.Column(db.String(), unique=False)
     phone_number = db.Column(db.String(), unique=True)
     profile_picture = db.Column(db.String())
     
@@ -174,8 +174,7 @@ class Landlords(db.Model):
                 'birth_date': self.birth_date,
                 'dni': self.dni,
                 'phone_number': self.phone_number,
-                'profile_picture': self.profile_picture,
-                'id': self.id}
+                'profile_picture': self.profile_picture}
 
 
 class Flats(db.Model):
