+++ src/front/js/component/Navbar.jsx
@@ -7,6 +7,9 @@ import "../../styles/navbar.css";
 
 export const Navbar = () => {
 	const { store, actions } = useContext(Context);
+	const handleLogout = () => {
+		actions.userLogout()
+	}
 
 	return (
 		<nav className="navbar-custom">
@@ -19,10 +22,9 @@ export const Navbar = () => {
 				<Link to="/demo">
 					<span>Lorem</span>
 				</Link>
-
 				<Link to="/uploadimg">
 					<span>Upload</span>
-        </Link>
+				</Link>
 				<Link to="/uploadflat">
 					<span>Upload Flat</span>
 				</Link>
@@ -30,13 +32,42 @@ export const Navbar = () => {
 					<span>Chats</span>
 				</Link>
 				{store.isLogedIn ?
-					<span>
-						<Link to="/dashboard">
-							<span className="red-color">Perfil</span>
-							<span className="profile-picutre"><img src={profilePicture} /></span>
-						</Link>
-					</span> 
-					: 
+					<div class="dropdown">
+						<button class="btn btn-custom red-background dropdown-toggle" type="button" data-bs-toggle="dropdown" aria-expanded="false">
+							Menú
+							<span className="profile-picutre"><img src={store.userData.profile_picture} /></span>
+						</button>
+						<ul class="dropdown-menu">
+							<li><a class="dropdown-item" href="#">
+								<Link to="/dashboard">
+									<span className="red-color">Perfil</span>
+								</Link>
+							</a></li>
+							<li><a class="dropdown-item" href="#">
+								{store.userData.is_student ?
+									<Link to="/favorites">
+										<span className="red-color">Favoritos</span>
+									</Link>
+									:
+									<Link to="/myflats">
+										<span className="red-color">Mis pisos</span>
+									</Link>
+								}
+							</a></li>
+							<li><a class="dropdown-item" href="#">
+								<Link to="/chats/1">
+									<span className="red-color">Mensajes</span>
+								</Link>
+							</a></li>
+							<li><a class="dropdown-item" href="#">
+								<Link to="/">
+									<button className="btn-custom red-background" onClick={handleLogout}> <i className="fas fa-sign-out-alt"></i> Desconectar</button>
+								</Link>
+							</a></li>
+						</ul>
+					</div>
+
+					:
 					<>
 						<Link to="/login">
 							<button className="btn-custom red-background">Login</button>
